requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.0.0
pygame==2.6.1
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
torch==2.8.0+cpu
torchvision==0.23.0+cpu
typing_extensions==4.12.2
